[SQS] Print access_key when SQS auth failed.
boto will try to get access_key and secret_key from many sources:

1. get from arguments.
2. get from .boto configuration file.
3. get from environment variables.
4. get from metadata service in elastic beanstalk environment.

So when an access denied error occurrs, it is hard to identify whether we're
using the wrong account or if the access policy is wrong.

This fix will show the access_key for access denied error, which will
help developer diagnose the problem.