Add Consul K/V store Transport
Using the Key/Value store of Consul.io this commit adds a new
Transport.

It is a simple Transport which Messages as Keys in the K/V store
of Consul using the HTTP API.

The HTTP API is consumed using the python-consul client which exposes
the API in a Python way.